refactor(blackbox): use firmware's conversions.cpp for unit tests
Use the actual firmware conversion functions instead of duplicated
inline implementations in the blackbox test render mock.

Changes:
- platformio.ini: Include src/conversions.cpp in blackbox build
- render_mock.cpp/h: Remove duplicate conversion functions, use
  celsius_to_fahrenheit(), meterspersecond_to_milesperhour(),
  hectopascals_to_millimetersofmercury(), etc. from firmware
- api_response.cpp: Remove duplicate RenderEvent::toJson()
- mock_server/main.py: Fix sunny scenario weather_code to 0 (clear sky)

This ensures unit conversions in tests exactly match the firmware.

All 10 blackbox tests pass.

+++ platformio/test/test_blackbox/render_mock.h
@@ -164,7 +164,7 @@ public:
     /**
      * Get the drawn humidity value (0 if not found)
      */
-    int getDrawnHumidity() const;
+    float getDrawnHumidity() const;
     
     /**
      * Get the drawn wind speed value (0 if not found)
